fixing tests; reorganizing stubs into their own folder
