--- tests/Hangfire.Core.Tests/Dashboard/UnauthorizedDashboardFilter.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using Hangfire.Dashboard;
-
-namespace Hangfire.Core.Tests.Dashboard
-{
-    internal class UnauthorizedDashboardFilter : IDashboardAuthorizationFilter
-    {
-        public bool Authorize(DashboardContext context)
-        {
-            return false;
-        }
-    }
-}
