+++ tests/Hangfire.Core.Tests/Stubs/AuthorizedDashboardFilterStub.cs
@@ -1,0 +1,12 @@
+﻿using Hangfire.Dashboard;
+
+namespace Hangfire.Core.Tests.Stubs
+{
+    internal class TestAuthorizedDashboardFilter : IDashboardAuthorizationFilter
+    {
+        public bool Authorize(DashboardContext context)
+        {
+            return true;
+        }
+    }
+}
