+++ tests/Hangfire.Core.Tests/Stubs/DashboardContextStub.cs
@@ -1,0 +1,12 @@
+﻿using Hangfire.Dashboard;
+
+namespace Hangfire.Core.Tests.Stubs
+{
+    class DashboardContextStub : DashboardContext
+    {
+        public DashboardContextStub(DashboardOptions options) : base(new JobStorageStub(), options)
+        {
+            Response = new DashboardResponseStub();
+        }
+    }
+}
