+++ tests/Hangfire.Core.Tests/Stubs/UnauthorizedDashboardFilterStub.cs
@@ -1,0 +1,12 @@
+﻿using Hangfire.Dashboard;
+
+namespace Hangfire.Core.Tests.Stubs
+{
+    internal class UnauthorizedDashboardFilterStub : IDashboardAuthorizationFilter
+    {
+        public bool Authorize(DashboardContext context)
+        {
+            return false;
+        }
+    }
+}
